+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 FaasJS use [Semantic Versioning](https://semver.org/).
 
+Unreleased
+
+- `@faasjs/pg`
+  - [Feature] Add `faasjs-pg sql` to execute SQL from arguments or stdin using `DATABASE_URL`.
+
 [`v8.0.0-beta.36 (2026-06-05)`](https://github.com/faasjs/faasjs/compare/v8.0.0-beta.35...v8.0.0-beta.36)
 
 - `faasjs`